rename tenantattr to volumetenants; remove commented code
